+++ CHANGELOG.rst
@@ -1,5 +1,9 @@
+ 0.2.0 (2015-04-20)
+  * added option to get top-n longest sequential patterns (option "--top n")
+
  0.1.2
   * enhanced calculating of difference between sequences
+  * improved procedure for processing of 1- and 2-sequences
 
  0.1.1 (2015-03-31)
   * doc-strings updates